requirements.txt
langchain==0.0.331
python-dotenv==1.0.0
streamlit==1.27.2
